Switch Docker instructions to always `--pull` on build.
This fixes situations like Vincent hit, where a stale base image would lead to
new packages based on old base packages.
Change: 118071412

+++ tensorflow/python/ops/control_flow_ops.py
@@ -897,7 +897,7 @@ def ZerosLikeOutsideLoop(op, index):
     branch = op_ctxt.branch
     switch_val = switch(op.inputs[0], pred)[1 - branch]
     zeros_shape = array_ops.shape(switch_val)
-    return array_ops.zeros(zeros_shape)
+    return array_ops.zeros(zeros_shape, dtype=val.dtype)
 
 
 class ControlFlowContext(object):
